NB-RU-06: Testimonial Completed (#8)
* NB-RU-06: Testimonial completed

* NB-RU-06: Color redundancy reduced

* NB-RU-06: Arrows fixed

* NB-RU-06: Review fixes

* NB-RU-06: Testimonials review fixes
