+++ index.html
@@ -359,6 +359,142 @@
             </div>
         </section>
 
+        <!-- Testimonial Section -->
+        <section
+            class="testimonial container"
+            aria-labelledby="testimonial-title"
+        >
+            <div class="testimonial-background" aria-hidden="true"></div>
+            <header class="testimonial__content-details">
+                <h2 class="testimonial__content-title" id="testimonial-title">
+                    Testimonial
+                </h2>
+                <h3 class="testimonial__content-description">
+                    Trust our clients
+                </h3>
+            </header>
+            <div class="splide">
+                <div class="splide__arrows testimonial-arrows">
+                    <button
+                        class="splide__arrow splide__arrow--prev testimonial-prev"
+                    >
+                        <i class="icon-arrow-left"></i>
+                    </button>
+                    <button
+                        class="splide__arrow splide__arrow--next testimonial-next"
+                    >
+                        <i class="icon-arrow-right"></i>
+                    </button>
+                </div>
+
+                <div class="splide__track">
+                    <ul class="splide__list">
+                        <!-- Slide 1 -->
+                        <li class="splide__slide">
+                            <div class="testimonial-avatar">
+                                <img
+                                    src="./src/assets/images/admin.png"
+                                    alt="Testimonial Avatar"
+                                />
+                            </div>
+                            <div class="testimonial__details">
+                                <div class="testimonial__review">
+                                    <p class="testimonial__review-user">
+                                        <span>Mark Smith</span> / Travel
+                                        Enthusiast
+                                    </p>
+                                    <div
+                                        class="testimonial__review-ratings"
+                                        aria-hidden="true"
+                                    >
+                                        <i class="icon-star"></i>
+                                        <i class="icon-star"></i>
+                                        <i class="icon-star"></i>
+                                        <i class="icon-star"></i>
+                                        <i class="icon-star"></i>
+                                    </div>
+                                </div>
+                                <p class="testimonial__description">
+                                    Contrary to popular belief, Lorem Ipsum is
+                                    not simply random text. It has roots in a
+                                    piece of classical Latin literature from 45
+                                    BC.
+                                </p>
+                            </div>
+                        </li>
+
+                        <!-- Slide 2 -->
+                        <li class="splide__slide">
+                            <div class="testimonial-avatar">
+                                <img
+                                    src="./src/assets/images/admin.png"
+                                    alt="Testimonial Avatar"
+                                />
+                            </div>
+                            <div class="testimonial__details">
+                                <div class="testimonial__review">
+                                    <p class="testimonial__review-user">
+                                        <span>Mark Smith</span> / Travel
+                                        Enthusiast
+                                    </p>
+                                    <div
+                                        class="testimonial__review-ratings"
+                                        aria-hidden="true"
+                                    >
+                                        <i class="icon-star"></i>
+                                        <i class="icon-star"></i>
+                                        <i class="icon-star"></i>
+                                        <i class="icon-star"></i>
+                                        <i class="icon-star"></i>
+                                    </div>
+                                </div>
+                                <p class="testimonial__description">
+                                    Contrary to popular belief, Lorem Ipsum is
+                                    not simply random text. It has roots in a
+                                    piece of classical Latin literature from 45
+                                    BC.
+                                </p>
+                            </div>
+                        </li>
+
+                        <!-- Slide 3 -->
+                        <li class="splide__slide">
+                            <div class="testimonial-avatar">
+                                <img
+                                    src="./src/assets/images/admin.png"
+                                    alt="Testimonial Avatar"
+                                />
+                            </div>
+                            <div class="testimonial__details">
+                                <div class="testimonial__review">
+                                    <p class="testimonial__review-user">
+                                        <span>Mark Smith</span> / Travel
+                                        Enthusiast
+                                    </p>
+                                    <div
+                                        class="testimonial__review-ratings"
+                                        aria-hidden="true"
+                                    >
+                                        <i class="icon-star"></i>
+                                        <i class="icon-star"></i>
+                                        <i class="icon-star"></i>
+                                        <i class="icon-star"></i>
+                                        <i class="icon-star"></i>
+                                    </div>
+                                </div>
+                                <p class="testimonial__description">
+                                    Contrary to popular belief, Lorem Ipsum is
+                                    not simply random text. It has roots in a
+                                    piece of classical Latin literature from 45
+                                    BC.
+                                </p>
+                            </div>
+                        </li>
+                    </ul>
+                </div>
+            </div>
+        </section>
+
         <script type="module" src="./src/script/index.js"></script>
     </body>
 </html>
