+++ src/styles/base/_typography.scss
@@ -53,7 +53,6 @@ h3 {
     color: map-get(variables.$colors, grey-black);
     line-height: 1.2;
     text-align: center;
-    color: map-get($map: variables.$colors, $key: grey-black);
 
     @include breakpoint('lg') {
         font-size: px-to-rem(40);
