+++ src/styles/layout/_header.scss
@@ -1,7 +1,6 @@
 @use '../abstracts/functions' as *;
 @use '../abstracts/variables' as variables;
 @use '../abstracts/mixins' as *;
-@use '../base/typography';
 
 // ================= BASE =================
