+++ src/styles/layout/_index.scss
@@ -4,3 +4,4 @@
 @forward 'services';
 @forward 'travelpoint';
 @forward 'keyfeatures';
+@forward 'testimonial';
