+++ src/styles/layout/_services.scss
@@ -1,7 +1,6 @@
 @use '../abstracts/functions' as *;
 @use '../abstracts/variables' as variables;
 @use '../abstracts/mixins' as *;
-@use '../base/typography';
 
 .services {
     @include flex($direction: column, $gap: px-to-rem(32));
