+++ src/styles/layout/_testimonial.scss
@@ -1,0 +1,212 @@
+@use '../abstracts/functions' as *;
+@use '../abstracts/variables' as variables;
+@use '../abstracts/mixins' as *;
+
+.testimonial {
+    position: relative;
+    padding: px-to-rem(70) px-to-rem(16);
+    margin-top: px-to-rem(-30);
+    @include flex($direction: column, $gap: px-to-rem(32));
+
+    @include breakpoint('lg') {
+        max-width: 100%;
+        margin-top: px-to-rem(26);
+        padding: px-to-rem(52) px-to-rem(32);
+        gap: px-to-rem(27);
+    }
+
+    &-background {
+        @include breakpoint('lg') {
+            position: absolute;
+            top: 0;
+            left: 0;
+            width: 100%;
+            height: 100%;
+            background-image: url(/src/assets/images/testimonial-background.png);
+            background-size: cover;
+            background-position: center;
+            z-index: -1;
+        }
+
+        @include breakpoint('xl') {
+            top: 15%;
+        }
+    }
+
+    @include breakpoint('xl') {
+        max-width: px-to-rem(1440);
+        margin-top: px-to-rem(-66);
+        gap: px-to-rem(26);
+        padding: px-to-rem(68);
+    }
+
+    &__details {
+        @include flex($direction: column, $gap: px-to-rem(66));
+
+        @include breakpoint('lg') {
+            gap: px-to-rem(57);
+        }
+    }
+
+    &__review {
+        @include flex($direction: column, $gap: px-to-rem(29));
+
+        &-user {
+            font-size: px-to-rem(23);
+            font-weight: bold;
+        }
+
+        span {
+            color: map-get($map: variables.$colors, $key: orange);
+            font-size: px-to-rem(28);
+        }
+
+        &-ratings {
+            @include flex($gap: px-to-rem(16));
+
+            i {
+                color: map-get($map: variables.$colors, $key: yellow);
+                font-size: px-to-rem(30);
+            }
+        }
+    }
+
+    &__description {
+        font-size: px-to-rem(17);
+
+        @include breakpoint('lg') {
+            max-width: 65%;
+            font-size: px-to-rem(22);
+        }
+
+        @include breakpoint('xl') {
+            max-width: 62%;
+            text-align: center;
+            line-height: 1.7;
+        }
+    }
+
+    &__content {
+        @include flex($direction: column, $gap: px-to-rem(32));
+
+        &-details {
+            @include flex($direction: column, $gap: px-to-rem(16));
+
+            @include breakpoint('xl') {
+                gap: px-to-rem(16);
+            }
+        }
+
+        @include breakpoint('lg') {
+            gap: px-to-rem(32);
+        }
+
+        @include breakpoint('xl') {
+            align-items: flex-start;
+            gap: px-to-rem(30);
+        }
+
+        &-title {
+            @include breakpoint('lg') {
+                letter-spacing: px-to-rem(6);
+            }
+        }
+    }
+
+    &__pagination {
+        @include flex($justify: center, $gap: px-to-rem(25));
+        margin-top: px-to-rem(68);
+        margin-bottom: px-to-rem(-23);
+
+        &-page {
+            background-color: color-opacity(
+                map-get(variables.$colors, grey-black),
+                0.05
+            );
+            padding: px-to-rem(12);
+            border-radius: 50%;
+
+            &.is-active {
+                background-color: map-get(variables.$colors, secondary);
+            }
+        }
+
+        @include breakpoint('lg') {
+            margin-top: px-to-rem(57);
+        }
+
+        @include breakpoint('xl') {
+            margin-top: px-to-rem(53);
+        }
+    }
+}
+
+.splide {
+    width: 100%;
+    &__list {
+        transform: translateX(px-to-rem(-1921));
+
+        @include breakpoint('xl') {
+            transform: translateX(px-to-rem(-5216));
+        }
+    }
+    &__track {
+        cursor: grab;
+        max-width: 100%;
+        margin-top: px-to-rem(29);
+
+        &:active {
+            cursor: grabbing;
+        }
+    }
+
+    &__slide {
+        @include flex($direction: column, $gap: px-to-rem(57));
+
+        @include breakpoint('lg') {
+            gap: px-to-rem(51);
+        }
+    }
+
+    &__arrow {
+        top: 119%;
+        padding: unset;
+        font-size: 10px;
+        padding: px-to-rem(32);
+        color: map-get($map: variables.$colors, $key: grey-black);
+        background-color: map-get($map: variables.$colors, $key: light);
+
+        &:hover {
+            background-color: map-get($map: variables.$colors, $key: primary);
+        }
+
+        @include breakpoint('lg') {
+            top: 40%;
+            padding: px-to-rem(50);
+        }
+
+        &--prev {
+            left: px-to-rem(0);
+
+            @include breakpoint('lg') {
+                left: px-to-rem(32);
+            }
+
+            @include breakpoint('xl') {
+                left: px-to-rem(54);
+            }
+        }
+
+        &--next {
+            right: px-to-rem(0);
+
+            @include breakpoint('lg') {
+                right: px-to-rem(32);
+            }
+
+            @include breakpoint('xl') {
+                right: px-to-rem(54);
+            }
+        }
+    }
+}
