+++ src/styles/vendors/_index.scss
@@ -1,0 +1,2 @@
+@forward 'icomoon';
+@forward 'splide';
